+++ sr_fingertip_visualization/launch/tactile_visualizer.launch
@@ -1,0 +1,5 @@
+<launch>
+  <arg name="rosbag_path" default=""/>
+  <node pkg="sr_fingertip_visualization" type="sr_fingertip_visualization_gui.py" name="fingertip_gui" output="screen"/>
+  <node pkg="rosbag" type="play" name="rosbag" args="$(arg rosbag_path) -l"  unless="$(eval arg('rosbag_path') == '')"/>
+</launch>
