+++ sr_fingertip_visualization/resource/hand.png
