+++ sr_fingertip_visualization/scripts/sr_fingertip_plugin_gui
