+++ sr_fingertip_visualization/src/sr_fingertip_visualization/finger_widgets_visual.py
@@ -32,12 +32,12 @@ from python_qt_binding.QtWidgets import (
     QStackedLayout
 )
 
-from sr_gui_fingertip_visualization.tactile_points import (
+from sr_fingertip_visualization.tactile_points import (
     TactilePointPST,
     TactilePointBiotacSPPlus,
     TactilePointBiotacSPMinus
 )
-from sr_gui_fingertip_visualization.tab_layouts_generic import GenericTabLayout
+from sr_fingertip_visualization.tab_layouts_generic import GenericTabLayout
 from sr_robot_msgs.msg import ShadowPST, BiotacAll
