+++ sr_fingertip_visualization/src/sr_fingertip_visualization/tab_layouts_generic.py
