+++ sr_fingertip_visualization/src/sr_fingertip_visualization/tactile_point_generic.py
