+++ sr_fingertip_visualization/src/sr_fingertip_visualization/tactile_points.py
