--- sr_gui_fingertip_visualization/launch/tactile_visualizer.launch
@@ -1,3 +1,0 @@
-<launch>
-  <node pkg="sr_data_visualization" type="sr_gui_fingertip_visualization.py" name="fingertip_gui" output="screen"/>
-</launch>
